requirements_dev.txt
httpie==3.2.1
django-extensions==3.2.0
ipython==8.4.0